fix(typescript): Fix type definition for typescript command definition (#8)
BUG
Some tests for the multi command using typescript were failing. This was because of a typescript error:

src/commands/hello.ts(12,10): error TS2742: The inferred type of 'flags' cannot be named without a reference to 'alto-command/node_modules/@oclif/parser/lib/flags'. This is likely not portable. A type annotation is necessary.

DETAILS
This is a baffling error. If I replaced the reference to alto-command with @oclif/command, the error would vanish. The two packages are essentially identical. Perhaps there's some indirect reference to @oclif/command in one of the dependencies and so things aren't quite meshing.

However, as I investigated further, I found that the types in a pure oclif system do not work correctly. Consider

    const foo: flags.IFlag<string> = {. // where flags is the import from @oclif/command
      type: 'boolean',
      allowNo: false,
      name: 'fred',
      parse: (input: boolean, context: any) => { return 'hi' }
    };

This is, according to typescript a reasonable declaration. However, I can take the exact same thing and put into the static flags:

  static flags = {
    foo4: {
      type: 'boolean',
      allowNo: false,
      name: 'fred',
      parse: (input: boolean, context: any) => { return 'hi' }
    },

an I get another typescript error about properties not being correct. Which I believe is not right. Alternately, you can take the the output from something like flags.boolean({char: 'f'}) and insert that into the static flags declaration, and you'll get another error (even though, obviously, that should be correct). This suggests to me that the types around this static flags declaration has some problems. Note that the declaration of static flags in @oclif/command is static flags?: flags.Input<any>;, whee that any is pretty strange (since that is used to drive some of the type parameterization in the Input and Output types). Again, it currently seems like the various types here are not declared quite right in oclif.

FIX
This "fixes" the problem by modifying the typescript template to make the type of static flags explicit (static flags: flags.Input<any> = ...). I don't find this satisfying, but I can't justify the time to untangle and rewrite the types in oclif now.